refactor: organize files into usecase dirs
